Move build test into test_repository
The build test was mostly left here for historical purposes when we didn't have
pytest_container yet and had to ensure that the containers have been
built/pulled down beforehand. Now it makes no sense to have this test anymore,
thus we move the only useful test function into test_repository
