+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -6,7 +6,7 @@ run all test environments, add the following line to your PR description on a
 separate line! E.g.: [CI:TOXENVS] postgres,minimal
 
 The following tox environments are always added:
-build, all, repository, metadata, multistage
+all, repository, metadata, multistage
 
 You can obtain the list of all available environments by running `tox -l`.
 -->
